redirecting to new url

+++ main.py
@@ -126,12 +126,14 @@ def post_new_bistro():
 
 
 ## HTTP DELETE - Delete Record
-@app.route("/delete", methods=["DELETE", "GET"])
+@app.route("/delete/", methods=["DELETE", "GET"])
 def delete():
-    bistro_id = request.args.get("id")
-    bistro = Bistro.query.get(bistro_id)
-    db.session.delete(bistro)
-    db.session.commit()
+    api_key = request.args.get("api-key")
+    if api_key == "TopSecretAPIKey":
+        bistro_id = request.args.get("id")
+        bistro = Bistro.query.get(bistro_id)
+        db.session.delete(bistro)
+        db.session.commit()
     return redirect("/#coffee-shops")
